Move neighbors.DistanceMetric to metrics
The associated _typedefs.pyx file has been
moved to utils to avoid circular dependencies
has it is being used in neighbors.
