+++ sklearn/neighbors/_graph.py
@@ -65,10 +65,11 @@ def kneighbors_graph(
         between neighbors according to the given metric.
 
     metric : str, default='minkowski'
-        The distance metric used to calculate the k-Neighbors for each sample
-        point. The DistanceMetric class gives a list of available metrics.
-        The default distance is 'euclidean' ('minkowski' metric with the p
-        param equal to 2.)
+        The distance metric used to calculate the neighbors within a
+        given radius for each sample point. The default distance is
+        'euclidean' ('minkowski' metric with the param equal to 2.)
+        See the documentation of :class:`metrics.DistanceMetric`
+        for a list of available metrics.
 
     p : int, default=2
         Power parameter for the Minkowski metric. When p = 1, this is
@@ -158,9 +159,10 @@ def radius_neighbors_graph(
 
     metric : str, default='minkowski'
         The distance metric used to calculate the neighbors within a
-        given radius for each sample point. The DistanceMetric class
-        gives a list of available metrics. The default distance is
+        given radius for each sample point. The default distance is
         'euclidean' ('minkowski' metric with the param equal to 2.)
+        See the documentation of :class:`metrics.DistanceMetric`
+        for a list of available metrics.
 
     p : int, default=2
         Power parameter for the Minkowski metric. When p = 1, this is
